requirements.txt
accelerate==0.33.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
elastic-transport==8.15.1
elasticsearch==8.15.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.0
imageio-ffmpeg==0.5.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
proglog==0.1.10
psutil==6.0.0
pycparser==2.22
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==72.2.0
soundfile==0.12.1
sympy==1.13.2
tbb==2021.13.1
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.5
transformers==4.41.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
